+++ src/model/WorksOn.java
@@ -1,12 +1,11 @@
 package model;
 
-import java.math.BigDecimal; //for BigDecimal (ideal for decimal(5,2) in MySQL)
 public class WorksOn {
     private String essn;
     private int pno;
-    private BigDecimal hours;
+    private double hours;
 
-    public WorksOn(String essn, int pno, BigDecimal hours){
+    public WorksOn(String essn, int pno, double hours){
         this.essn=essn;
         this.pno=pno;
         this.hours=hours;
@@ -14,9 +13,9 @@ public class WorksOn {
 
     public String getEssn() {return essn;}
     public int getPno() {return pno;}
-    public BigDecimal getHours() {return hours;}
+    public double getHours() {return hours;}
 
     public void setEssn(String essn) {this.essn=essn;}
     public void setPno(int pno) {this.pno=pno;}
-    public void setHours(BigDecimal hours) {this.hours=hours;}
+    public void setHours(double hours) {this.hours=hours;}
 }
